+++ html2canvas/src/misc/errorWithScreenshot.ts
@@ -1,16 +1,22 @@
-import {Injector, ValueProvider} from "@angular/core";
-import {DOCUMENT} from "@angular/common";
+import {Injector, ValueProvider, PLATFORM_ID} from "@angular/core";
+import {DOCUMENT, isPlatformBrowser} from "@angular/common";
 import {ErrorWithStack, AnglrExceptionExtender, ErrorWithScreenShot, ANGLR_EXCEPTION_EXTENDERS} from "@anglr/error-handling";
 import * as html2canvas from 'html2canvas';
 
 /**
  * Extends error with screenshot of current application state
- * @param injector Injector used for obtaining dependencies
- * @param error Error that should be extended
+ * @param injector - Injector used for obtaining dependencies
+ * @param error - Error that should be extended
  */
 export const errorWithScreenShotExtender: AnglrExceptionExtender = (injector: Injector, error: ErrorWithStack): Promise<ErrorWithStack> =>
 {
     let errorWithScreenshot: ErrorWithStack & ErrorWithScreenShot = error;
+
+    if(!isPlatformBrowser(injector.get(PLATFORM_ID)))
+    {
+        return Promise.resolve(errorWithScreenshot);
+    }
+
     let document = injector.get(DOCUMENT);    
 
     return new Promise(resolve =>