+++ src/exceptionHandling/errorWithUrlExtender.ts
@@ -6,8 +6,8 @@ import {AnglrExceptionExtender, ErrorWithUrl, ANGLR_EXCEPTION_EXTENDERS} from ".
 
 /**
  * Extends error with current application URL
- * @param injector Injector used for obtaining dependencies
- * @param error Error that should be extended
+ * @param injector - Injector used for obtaining dependencies
+ * @param error - Error that should be extended
  */
 export const errorWithUrlExtender: AnglrExceptionExtender = (injector: Injector, error: ErrorWithStack): Promise<ErrorWithStack> =>
 {